Add config handler to save memory values to memory
This way of handling configs allows commands and events to not have to
fetch the config values from file all the time, saving cpu resources,
while using a very minimal amount of memory in the process. This means
that the config can't be updated live, but an alternative solution will
be implemented later.