Add GUI menu for multi beds
This menu should display a list of beds you have set, alongside a
button to add new homes. This new home button opens up an anvil menu
which is used to set the name of the home you are setting. GUI menus
make use of the Inventory Framework which is set as a dependency.