Change repositories to focus on in memory storage
To ensure maximum speed, all homes and warps are stored in memory on
initialisation, with fetches obtained from the in memory representation
of the data. While this will result in increased memory usage, it is a
much needed tradeoff as memory is cheap while access times in a live
environment could be detrimental.