Add warpmenu command to test warp displaying
A warp menu can be utilised to view all the warps that currently exist
in the world. This will be changed to only show ones that you have
access to in the future once warp unlocking is implemented. This is
currently a test to show what the gui will potentially look like.