Add configurable back button on warp menu
By specifying a string on the warpmenu command, it changes the back
button from "exit" as in closing the inventory, to "go back", which runs
the set command. While the set command can be anything, it is intended
to be used to open a different menu, allowing simple linkage to other
menu systems.